backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
anthropic==0.40.0
pdfplumber==0.11.4
python-docx==1.1.2
reportlab==4.2.5
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.0
beautifulsoup4==4.12.3
